==> notebooks/weather/0_get_weather_station.ipynb <==
# Get Weather Station Metadata (Synoptic API)

<b>Summary:</b><br>
Uses Synoptic Weather Data API to get weather station metadata.<br>
Synoptic is partner of SDG&E, helping store and serve their weather station data--accessible via [API](https://developers.synopticdata.com/mesonet/).

- Get weather station metadata via synoptic API
- Save raw weather station metadata file

<b>Output:</b><br>
.<br>
└── data<br>
&emsp;&emsp;&emsp;└── raw<br>
&emsp;&emsp;&emsp;&emsp;&emsp;&emsp;└── station_metadata.csv

<b>Instructions:</b><br>
- See README.md in directory

<b>Areas for Improvement:</b><br>

import json

import pandas as pd
import requests

## Parameters

# Base url and API token
with open("config.json") as config_file:
    cfg = json.load(config_file)

API_BASE_URL = cfg["synoptic_base_url"]
API_TOKEN = cfg["synoptic_api_token"]

# should be SHORTNAME values from mesonet
WEATHER_NETWORKS = cfg["weather_networks"]
WEATHER_NETWORKS = {int(k): v for k, v in WEATHER_NETWORKS.items()}  # convert id to int

## Get all weather station metadata for the given networks

### Helper function to pull weather station metadata using synoptic API

def get_station_metadata(network: str) -> dict:
    """
    Return json dictionary of station metadata within given network(s).
    """
    endpoint = f"{API_BASE_URL}stations/metadata"
    params = {
        "token": API_TOKEN,
        "network": network,
        "status": "ACTIVE",
        "obtimezone": "UTC",
        "sensorvars": 1,
        "complete": 1,
    }
    r = requests.get(endpoint, params=params)
    if r.status_code not in range(200, 300):
        # TODO: add error handling
        print("Error!")
        return {}
    return r.json()

### Get station metadata for each network

# convert keys to strings and commma separate
networks = ",".join([str(i) for i in list(WEATHER_NETWORKS.keys())])
stations_df = pd.json_normalize(get_station_metadata(network=networks)["STATION"])

stations_df

## Write weather station metadata to raw folder

stations_df.to_csv("../../data/raw/station_metadata.csv", index=False)
