==> notebooks/weather/2_get_and_process_weather_data.ipynb <==
# Get and Process Weather Data for Mapped Weather Stations

<b>Summary:</b><br>
Uses Synoptic Weather Data API to grab weather station data for the mapped weather stations to cameras.<br>
Synoptic is partner of SDG&E, helping store and serve their weather station data--accessible via API.

- Read in processed camera weather station mappings
- Get all station data for input timerange
- Save weather station data by network
**Note: This stores all weather data for a single network in a single dataframe before writing to output; depending on timeframe and number of stations, could be a large dataframe that exceeds computer memory. Computer used for execution contains 64GB memory.**
- Convert wind speed and direction to uv components
- Clean/snap interval reading times
- Save processed data with only desired columns by network

<b>Output:</b><br>
.<br>
├── data<br>
&emsp;&emsp;&emsp;├── processed<br>
&emsp;&emsp;&emsp;&nbsp;│&emsp;&emsp;&nbsp;├── weather_HPWREN.csv<br>
&emsp;&emsp;&emsp;&nbsp;│&emsp;&emsp;&nbsp;├── weather_SC-EDISON.csv<br>
&emsp;&emsp;&emsp;&nbsp;│&emsp;&emsp;&nbsp;├── weather_SDGE.csv<br>
&emsp;&emsp;&emsp;└── raw<br>
&emsp;&emsp;&emsp;&emsp;&emsp;&emsp;├── weather_HPWREN.csv<br>
&emsp;&emsp;&emsp;&emsp;&emsp;&emsp;├── weather_SC-EDISON.csv<br>
&emsp;&emsp;&emsp;&emsp;&emsp;&emsp;└── weather_SDGE.csv

<b>Instructions:</b><br>
- See README.md in directory

<b>Areas for Improvement:</b><br>
- Switch back to storing weather data files by station to reduce file size for any client running this
- Make get weather data call async to improve speed and/or multi process
- Cleanup imports

import asyncio
import json
import os
import time
import urllib
from ast import literal_eval
from datetime import datetime, timedelta
from io import StringIO

import aiofiles
import aiohttp
import fastparquet
import geopandas
import matplotlib.pyplot as plt
import nest_asyncio
import numpy as np
import pandas as pd
import pytz
import requests
import rtree
from aiohttp import ClientSession
from geopandas import GeoDataFrame
from shapely.geometry import Point, Polygon
from tqdm.notebook import tqdm, trange

nest_asyncio.apply()

## Parameters

# Base url and API token
with open("config.json") as config_file:
    cfg = json.load(config_file)

API_BASE_URL = cfg["synoptic_base_url"]
API_TOKEN = cfg["synoptic_api_token"]

# should be SHORTNAME values from mesonet
WEATHER_NETWORKS = cfg["weather_networks"]
WEATHER_NETWORKS = {int(k): v for k, v in WEATHER_NETWORKS.items()}  # convert id to int

# figlib start = 2016-06-04 (assume PST)
# figlib end = 2021-07-11 (assume PST)
# YYYYmmddHHMM (format needed for api)
START_DATETIME = cfg["weather_start_datetime"]
END_DATETIME = cfg["weather_end_datetime"]

## 1. Get data for matched weather stations

### Read weather station mappings

station_mappings_df = pd.read_csv("../../data/processed/camera_station_mappings.csv")
print(station_mappings_df.shape)

station_mappings_df

### Helper function to get weather station data given stid and start/end time

def get_historical_station_readings(stid: str, start: str, end: str) -> str:
    """
    Return csv string of station readings given station id(s) and timeframe.
    Start and end format = YYYYmmddHHMM. Synchronous.

    Note:
    For multiple stids, would need to remove output csv, but would be in json format.
    """
    # All times are requested in UTC, but may be returned in either UTC or Local time
    endpoint = f"{API_BASE_URL}stations/timeseries"
    params = {
        "token": API_TOKEN,
        "stid": stid,
        "start": start,
        "end": end,
        "obtimezone": "UTC",
        "output": "csv",
    }
    r = requests.get(endpoint, params=params)
    if r.status_code not in range(200, 300):
        # TODO: add error handling
        print("Error!")
        return {}
    # return r.json()
    return r.text

def get_network_station_data(df: pd.DataFrame, shortname: str = None) -> pd.DataFrame:
    """
    Return dataframe of station data given weather network shortname.
    If no shortname given, go through the full dataframe.
    Use if getting data in single file or in files by network.
    """
    if shortname != None:
        df = df[df["shortname"] == shortname]
    stations = df["stid"].dropna().unique().tolist()

    weather_df_created = 0

    for idx, station in enumerate(tqdm(stations)):
        # print(station)
        # get csv string
        try:
            csv_string_io = StringIO(
                get_historical_station_readings(
                    stid=station,
                    start=START_DATETIME,
                    end=END_DATETIME
                    # start="201606030000",
                    # end="201606030100",
                )
            )
            # initial creation of df
            if not weather_df_created:
                weather_df = pd.read_csv(csv_string_io, skiprows=6, header=[0, 1])
                weather_df_created = 1
            # additional appends to df
            else:
                temp_df = pd.read_csv(csv_string_io, skiprows=6, header=[0, 1])
                weather_df = pd.concat([weather_df, temp_df])
        except:
            print(f"Error for on network {shortname} stid: {station}")

    return weather_df

async def get_station_data(
    stid: str, start: str, end: str, session: ClientSession
) -> str:
    """
    INCOMPLETE.
    Return weather station csv string data given stid. Asynchronous.
    Start and end format = YYYYmmddHHMM.
    """
    # All times are requested in UTC, but may be returned in either UTC or Local time
    endpoint = f"{API_BASE_URL}stations/timeseries"
    params = {
        "token": API_TOKEN,
        "stid": stid,
        "start": start,
        "end": end,
        "obtimezone": "UTC",
        "output": "csv",
    }
    params_encode = urllib.parse.urlencode(params)
    url = endpoint + "?" + params_encode
    print(url)
    # r = requests.get(endpoint, params=params)
    r = await session.request(method="GET", url=url)
    # r.raise_for_status()
    if r.status not in range(200, 300):
        # TODO: add error handling
        print(f"Error! - {stid}")
        return ""
    # return r.json()
    # return r.text
    r_text = await r.text()
    return r_text


async def save_station_data(
    stid: str, start: str, end: str, session: ClientSession
) -> None:
    """
    Get and save weather station data to csv. Asynchronous.
    Start and end format = YYYYmmddHHMM.
    """
    base_output_path = "../../data/raw/stations/"
    r_text = await get_station_data(stid, start, end, session)
    r_io = StringIO(r_text)
    r_df = pd.read_csv(r_io, skiprows=6, header=[0, 1])
    r_df.to_csv(f"{base_output_path}{stid}.csv", index=False)

# %%time

# # Synchronous baseline test

# test_stations = ["BFDSD"]
# test_stations = ["BFDSD", "CVXSD"]
# test_stations = ["BFDSD", "CVXSD", "DJZSD", "HP016", "MGDSD"]

# for station in test_stations:
#     # get_historical_station_readings(
#     #     stid=station, start=START_DATETIME, end=END_DATETIME
#     # )
#     tmp = StringIO(
#         get_historical_station_readings(
#             stid=station, start=START_DATETIME, end=END_DATETIME
#         )
#     )
#     tmp_df = pd.read_csv(tmp, skiprows=6, header=[0, 1])
#     tmp_df.to_csv(f"../../data/raw/stations0/{station}.csv", index=False)

# Asynchronous

# Current issue: timeout after ~50 stations; believe surpassing max requests allowed at once with apis

# stations = station_mappings_df["stid"].dropna().unique().tolist()

# start = time.time()

# my_conn = aiohttp.TCPConnector(limit=10)
# async with ClientSession(connector=my_conn) as session:
#     await asyncio.gather(
#         *[
#             save_station_data(station, START_DATETIME, END_DATETIME, session)
#             for station in stations
#         ]
#     )
# print(time.time() - start)

# Get for all stations grouped in files by network

for key in WEATHER_NETWORKS:
    print(f"key: {key}")
    network_station_df = get_network_station_data(
        station_mappings_df, WEATHER_NETWORKS[key]
    )
    network_station_df.to_csv(
        f"../../data/raw/weather_{WEATHER_NETWORKS[key]}.csv", index=False
    )

# %%time

# Get for all stations as single file

# weather_df_created = 0

# for idx, station in enumerate(tqdm(stations)):
#     # print(station)
#     # get csv string
#     csv_string_io = StringIO(
#         get_historical_station_readings(
#             stid=station,
#             start=START_DATETIME,
#             end=END_DATETIME
#             # start="201606030000",
#             # end="201606030100",
#         )
#     )
#     # initial creation of df
#     if not weather_df_created:
#         weather_df = pd.read_csv(csv_string_io, skiprows=6, header=[0, 1])
#         weather_df_created = 1
#     # additional appends to df
#     else:
#         temp_df = pd.read_csv(csv_string_io, skiprows=6, header=[0, 1])
#         weather_df = pd.concat([weather_df, temp_df])

# # prev runtime: 1h 6min 56s

<hr>

## 2. Convert wind speed and direction to uv components

Reference: http://colaweb.gmu.edu/dev/clim301/lectures/wind/wind-uv

Directions are originally in "weather wind direction" and will be converted to "math wind direction" for uv componenent calculation.<br>
Null wind speed will remain as null for vector component. Null wind direction results in 0 v component.

### Read raw weather data

weather_sdge_df = pd.read_csv("../../data/raw/weather_SDGE.csv", header=[0, 1])
weather_hpwren_df = pd.read_csv("../../data/raw/weather_HPWREN.csv", header=[0, 1])
weather_sce_df = pd.read_csv("../../data/raw/weather_SC-EDISON.csv", header=[0, 1])

### Helper function to calculate uv components

def calc_uv_components(df: pd.DataFrame) -> pd.DataFrame:
    """
    Return original dataframe with uv column componenets.
    """
    # convert direction to math direction
    df["wind_direction_math"] = 270 - df["wind_direction_set_1"]["Degrees"]
    # if negative add 360
    df.loc[df["wind_direction_math"] < 0, ["wind_direction_math"]] += 360
    # convert degrees to radians
    df["wind_direction_math_r"] = np.radians(df["wind_direction_math"])

    # calculate uv components
    df["u"] = df["wind_speed_set_1"]["m/s"] * np.cos(df["wind_direction_math_r"])
    df["v"] = df["wind_speed_set_1"]["m/s"] * np.sin(df["wind_direction_math_r"])
    return df

weather_sdge_df = calc_uv_components(weather_sdge_df)
weather_hpwren_df = calc_uv_components(weather_hpwren_df)
weather_sce_df = calc_uv_components(weather_sce_df)

weather_sdge_df.head()

## 3. Clean weather datetimes / snap to 10 interval

### Set Date_Time column to datetime type

weather_sdge_df["Date_Time"] = pd.to_datetime(
    weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"]
)
weather_hpwren_df["Date_Time"] = pd.to_datetime(
    weather_hpwren_df["Date_Time"]["Unnamed: 1_level_1"]
)
weather_sce_df["Date_Time"] = pd.to_datetime(
    weather_sce_df["Date_Time"]["Unnamed: 1_level_1"]
)

print(weather_sdge_df["Date_Time"].dtypes)
print(weather_hpwren_df["Date_Time"].dtypes)
print(weather_sce_df["Date_Time"].dtypes)

### Check interval counts not on 10 minute interval (ie 0...10...20...30...40...50)

def get_misaligned_datetime_count(df: pd.DataFrame) -> int:
    """
    Returns the rows count of datetimes with minutes not aligned to 10 minute interval given a dataframe.
    I.e. 0...10...20...30...40...50
    Assumes datetime column is ["Date_Time"]["Unnamed: 1_level_1"].
    """
    return df[(df["Date_Time"]["Unnamed: 1_level_1"].dt.minute % 10) != 0].shape[0]

print(
    f"SDGE misaligned datetime count: {get_misaligned_datetime_count(weather_sdge_df)}"
)
print(
    f"SDGE misaligned datetime percentage: {get_misaligned_datetime_count(weather_sdge_df)/weather_sdge_df.shape[0]*100}"
)

# Confirm if all seconds are 00
weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"].dt.second.value_counts()

print(
    f"HPWREN misaligned datetime count: {get_misaligned_datetime_count(weather_hpwren_df)}"
)
print(
    f"HPWREN misaligned datetime percentage: {get_misaligned_datetime_count(weather_hpwren_df)/weather_hpwren_df.shape[0]*100}"
)

# Confirm if all seconds are 00
weather_hpwren_df["Date_Time"]["Unnamed: 1_level_1"].dt.second.value_counts()

print(f"SCE misaligned datetime count: {get_misaligned_datetime_count(weather_sce_df)}")
print(
    f"SCE misaligned datetime percentage: {get_misaligned_datetime_count(weather_sce_df)/weather_sce_df.shape[0]*100}"
)

weather_sce_df["Date_Time"]["Unnamed: 1_level_1"].dt.second.value_counts()

Weather networks to fix datetimes within:

- SDGE
- HPWREN

### Clean SDGE

weather_sdge_df[
    (weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"].dt.minute % 10) != 0
]

#### Scenario: Additional timestamps added at random 5 min intervals

In this scenario the 10 minute intervals still exist, just some random mixed minute intervals seem to be sprinkled in. Drop these.

pd.set_option("display.max_rows", None)

weather_sdge_df[
    (
        weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"]
        >= datetime(2017, 7, 10, 14, 40, 0, 0, pytz.UTC)
    )
    & (
        weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"]
        < datetime(2017, 7, 10, 16, 0, 0, 0, pytz.UTC)
    )
    & (
        (weather_sdge_df["Station_ID"]["Unnamed: 0_level_1"] == "BFDSD")
        | (weather_sdge_df["Station_ID"]["Unnamed: 0_level_1"] == "CHOSD")
    )
]["Date_Time"]["Unnamed: 1_level_1"]

weather_sdge_df[
    (
        weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"]
        >= datetime(2019, 6, 3, 18, 0, 0, 0, pytz.UTC)
    )
    & (
        weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"]
        < datetime(2019, 6, 3, 18, 50, 0, 0, pytz.UTC)
    )
    & (weather_sdge_df["Station_ID"]["Unnamed: 0_level_1"] == "MLGSD")
]["Date_Time"]["Unnamed: 1_level_1"]

weather_sdge_df[
    (
        weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"]
        >= datetime(2017, 3, 9, 9, 0, 0, 0, pytz.UTC)
    )
    & (
        weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"]
        < datetime(2017, 3, 9, 10, 0, 0, 0, pytz.UTC)
    )
    & (weather_sdge_df["Station_ID"]["Unnamed: 0_level_1"] == "AMOSD")
]["Date_Time"]["Unnamed: 1_level_1"]

weather_sdge_df[
    (
        weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"]
        >= datetime(2020, 4, 23, 15, 0, 0, 0, pytz.UTC)
    )
    & (
        weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"]
        < datetime(2020, 4, 24, 0, 0, 0, 0, pytz.UTC)
    )
    & (weather_sdge_df["Station_ID"]["Unnamed: 0_level_1"] == "SHCSD")
]["Date_Time"]["Unnamed: 1_level_1"]

In this scenario, we actually see a temperary outage starting at at 15:30 fully resuming at 19:10.<br>
There is 1 mixed reading in between at 18:48. Drop this both for simplicity and because may it not be accurate during outage.

pd.reset_option("display.max_rows")

As reviewed above, we are okay to drop all the mixed intervals for the sdge weather network readings.

weather_sdge_df.shape

# Drop the 9 rows
index_list_to_drop = weather_sdge_df[
    (weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"].dt.minute % 10) != 0
].index.tolist()

weather_sdge_df.drop(index_list_to_drop, axis=0, inplace=True)

weather_sdge_df[
    (weather_sdge_df["Date_Time"]["Unnamed: 1_level_1"].dt.minute % 10) != 0
]

weather_sdge_df.shape

get_misaligned_datetime_count(weather_sdge_df)

### Clean HPWREN

weather_hpwren_df[
    (weather_hpwren_df["Date_Time"]["Unnamed: 1_level_1"].dt.minute % 10) != 0
].head(10)

weather_hpwren_df["Date_Time"]["Unnamed: 1_level_1"].dt.minute.value_counts()

#### Scenario: Mixed timestamps

HPWREN stations are not synced and are reading at mixed intervals that could start at 5, 6, 3, 4, 7, etc. <br>
We will need to snap these readings to their nearest 10 minute interval to properly aggregate with the other stations.<br>

See below for rounding:

Anything above 0, should round up to the next interval. In production, if the latest reading we have is from 7:03, that data will not exist for the 7:00 interval as it shouldn't be able to see into the future. Therefore, this number will be rounded up to be grouped with the next interval.

To achieve this, the ceiling of each timestamp to the nearest 10th minute will be used.

weather_hpwren_df["Date_Time_Snapped"] = weather_hpwren_df["Date_Time"][
    "Unnamed: 1_level_1"
].dt.ceil(freq="10T")
weather_hpwren_df[["Date_Time", "Date_Time_Snapped"]].head()

weather_hpwren_df["Date_Time"] = weather_hpwren_df["Date_Time_Snapped"]
weather_hpwren_df.drop(columns=["Date_Time_Snapped"], inplace=True)

weather_hpwren_df[
    (weather_hpwren_df["Date_Time"]["Unnamed: 1_level_1"].dt.minute % 10) != 0
].head(10)

weather_hpwren_df["Date_Time"]["Unnamed: 1_level_1"].dt.minute.value_counts()

get_misaligned_datetime_count(weather_hpwren_df)

### Write processed data to csv keeping only desired columns

columns = [
    "Station_ID",
    "Date_Time",
    "air_temp_set_1",
    "relative_humidity_set_1",
    "wind_speed_set_1",
    "wind_gust_set_1",
    "wind_direction_set_1",
    "dew_point_temperature_set_1d",
    "u",
    "v",
]

weather_sdge_df[columns].to_csv("../../data/processed/weather_SDGE.csv", index=False)
weather_hpwren_df[columns].to_csv(
    "../../data/processed/weather_HPWREN.csv", index=False
)
weather_sce_df[columns].to_csv(
    "../../data/processed/weather_SC-EDISON.csv", index=False
)
