==> notebooks/weather/3_process_weather_aggregation.ipynb <==
#importing all the required libaries 
import pandas as pd
import os
import os.path
import numpy as np
import shutil
import glob

### Import the path where the weather raw data is stored

path = os.getcwd()
path_add = os.path.normpath("../../data/processed/")
weather_data_file_path= os.path.normpath(path + os.sep + path_add)
weather_data_file_path

### Import the camera_to_weather_station mapping file

camera_station_mapping = pd.read_csv("../../data/processed/camera_station_mappings.csv")
camera_station_mapping.head()

camera_station_mapping.shape

### Import the camera ID to camera_name file (provided in Fiblib_files)

camera_station_mapping2 = pd.read_csv("../../data/processed/camera_image_id_mappings.csv")
camera_station_mapping2.rename(columns= { "image_id" : "camera_name"},inplace=True)
camera_station_mapping2 = camera_station_mapping2[['camera_id','camera_name']]
camera_station_mapping2.head()

### Merge both the mappnig files

camera_station_mapping3 = pd.merge(camera_station_mapping,camera_station_mapping2, how='left', left_on ="camera_id", right_on = "camera_id")

### Calculate inverse distance of every weather station so that closest station get higest weight

# Find max distance and subract from the distance column (to get inverse of distance)
max_distance_calculated = camera_station_mapping3.distance_mi.max()+1
camera_station_mapping3["distance_mi_inverse"] = max_distance_calculated - camera_station_mapping3["distance_mi"]
#camera_station_mapping3.head()

### Split the weather data into multiple small chuck of file.

class FileSettings(object):
    def __init__(self, file_name, row_size=1000000):
        self.file_name = file_name
        self.row_size = row_size


class FileSplitter(object):

    def __init__(self, file_settings):
        self.file_settings = file_settings

        if type(self.file_settings).__name__ != "FileSettings":
            raise Exception("Please pass correct instance ")

        self.df = pd.read_csv(self.file_settings.file_name,
                              chunksize=self.file_settings.row_size)

    def run(self, directory="split_files"):
        try:os.makedirs(directory)
        except Exception as e:pass

        counter = 0

        while True:
            try:
                file_name = "{}/{}_{}_row_{}.csv".format(
                    directory,  self.file_settings.file_name.split(".")[0], counter, self.file_settings.row_size
                ) 
                df = next(self.df).to_csv(file_name,index =False)
                counter = counter + 1
            except StopIteration:
                break
            except Exception as e:
                print("Error:",e)
                break

        return True

### Run the filter spliter for the 3 group of weather Stations
os.chdir(weather_data_file_path)

helper =  FileSplitter(FileSettings(
    file_name='weather_HPWREN.csv',
    row_size=2000000
))
helper.run()

helper =  FileSplitter(FileSettings(
    file_name='weather_SC-EDISON.csv',
    row_size=2000000
))
helper.run()

helper =  FileSplitter(FileSettings(
    file_name='weather_SDGE.csv',
    row_size=2000000
))
helper.run()

### Create the folder inside the temp directory weather_data by weatherstation directory

weather_data_file_path_split = weather_data_file_path + r'\split_files'
os.chdir(weather_data_file_path_split)
shutil.rmtree(weather_data_file_path_split +r'\ByWeatherStation',ignore_errors=True)

weather_split_file_list =  glob.glob('*.csv')
try:os.makedirs("ByWeatherStation")
except Exception as e:pass

weather_split_file_list

### split the weather data by station

def split_weather_station_by_file (file_weather_data):
    file_chunk_number = int(file_weather_data.split('_row_')[0].split("_")[-1])
    if(file_chunk_number == 0):
        temp_file1_df = pd.read_csv(file_weather_data,header=[0,1])
        temp_file1_df.columns = temp_file1_df.columns.droplevel(1)
    else:
        temp_file1_df = pd.read_csv(file_weather_data)
    
    weather_stid_in_temp_file = temp_file1_df.Station_ID.unique()
    for j in range(len(weather_stid_in_temp_file)):
        temp_weather_stid = weather_stid_in_temp_file[j]
        temp_file_iter = temp_file1_df[temp_file1_df.Station_ID == temp_weather_stid]
        if (temp_weather_stid+".csv" in os.listdir(r"ByWeatherStation/")):
            file_1 = pd.read_csv(r"ByWeatherStation/"+temp_weather_stid+".csv")
            file_1 = pd.concat([file_1,temp_file_iter])
            file_1.to_csv(r"ByWeatherStation/"+temp_weather_stid+".csv",index = False)
        else:
            temp_file_iter.to_csv(r"ByWeatherStation/"+temp_weather_stid+".csv", index = False)

for i in range(len(weather_split_file_list)):
    file_weather_data = weather_split_file_list [i]
    split_weather_station_by_file(file_weather_data)

weather_split_file_list[-1]

## Weather_data calculation by camera name

camera_name_unique_list = camera_station_mapping3.camera_name.unique()
len(camera_station_mapping3.camera_name.unique())

### create weather station folder by camera

camera_data_path = weather_data_file_path + r'\by_camera_summary'
os.chdir(weather_data_file_path)
shutil.rmtree(camera_data_path,ignore_errors=True)
os.mkdir("by_camera_summary")

os.chdir(weather_data_file_path + r"\split_files\ByWeatherStation")

###  remove the station id which donot have any weather_data

list_weather_station_data_csv = os.listdir()
weather_station_data_available = []
for i in range (len(list_weather_station_data_csv)):
    temp1= list_weather_station_data_csv[i].split(".")[0]
    weather_station_data_available.append(temp1)

camera_station_mapping3.shape

camera_station_mapping4 = camera_station_mapping3[camera_station_mapping3['stid'].isin(weather_station_data_available)]
#print(store_df_filtered)

camera_station_mapping4.shape

camera_station_mapping4.head()

### Loop through all weather stations to calculate the per camera summary

for i in range (len(camera_name_unique_list)):
    # Get the weather station related to camera
    camera_name = camera_name_unique_list[i]
    weather_station_of_intrest = list(camera_station_mapping4[camera_station_mapping4["camera_name"] == camera_name ].stid.unique())
    
    #read all the weather station data into single dataframe:
    #print(os.listdir())
    for j in range (len(weather_station_of_intrest)):    
        temp_df = pd.read_csv(weather_station_of_intrest[j] +".csv")
        if (j == 0):
            weather_station_of_intrest_df = temp_df.copy()
        else:
            weather_station_of_intrest_df =  pd.concat([weather_station_of_intrest_df,temp_df])
            
    # Stack all the weather data into single 
    weather_station_of_intrest_df_stacked = pd.melt(weather_station_of_intrest_df, id_vars=['Station_ID', 'Date_Time'],
        var_name='weather_parameter_name', value_name='weather_parameter_value')
    
    # extract key columns from the camera_station mapping file
    camera_station_mapping5 = camera_station_mapping4[camera_station_mapping4.camera_name == camera_name]
    camera_station_mapping5 = camera_station_mapping5 [["camera_name","distance_mi_inverse", "stid"]]
    
    # join the weather data to mapping file.
    weather_station_of_intrest_df_stacked = pd.merge(weather_station_of_intrest_df_stacked, camera_station_mapping5, left_on= "Station_ID", right_on= "stid")
  
    #aggregate the columns to find the sum of all the distance per time/weather_parameter.
    weather_station_of_intrest_df_pivot = pd.pivot_table (weather_station_of_intrest_df_stacked, values="distance_mi_inverse" , 
                                                index= ['Date_Time', 'weather_parameter_name'],
                                                 aggfunc='sum')
    weather_station_of_intrest_df_pivot.reset_index(inplace=True)
    weather_station_of_intrest_df_pivot.rename(columns = {'distance_mi_inverse': 'distance_mi_inverse_sum'}, inplace =True)
    
    
    #Join the weighted sum of distance to stacked columns and calculate the weight factor and normalize the weather data. 
    weather_station_of_intrest_df_stacked_with_weight = pd.merge(weather_station_of_intrest_df_stacked,weather_station_of_intrest_df_pivot, on= ['Date_Time', 'weather_parameter_name'])
    weather_station_of_intrest_df_stacked_with_weight['weather_parameter_value_norm'] =  (
    weather_station_of_intrest_df_stacked_with_weight['weather_parameter_value'] * weather_station_of_intrest_df_stacked_with_weight['distance_mi_inverse'] 
    / weather_station_of_intrest_df_stacked_with_weight['distance_mi_inverse_sum'])
    
    #Round data to nearest 3 digit. (run time is high)
    decimals = 4
    weather_station_of_intrest_df_stacked_with_weight['weather_parameter_value_norm'] = weather_station_of_intrest_df_stacked_with_weight['weather_parameter_value_norm'].apply(lambda x: round(x, decimals))
    
    # Calculate the weighted weather parameter value by per camera + Time. 
    weather_data_per_camera_agg = pd.pivot_table (weather_station_of_intrest_df_stacked_with_weight, values="weather_parameter_value_norm" , 
                                                index= ['Date_Time', 'camera_name'], columns= 'weather_parameter_name',
                                                 aggfunc='sum')
    weather_data_per_camera_agg.reset_index(inplace=True)
    
    weather_data_per_camera_agg_path = camera_data_path + "\\" + camera_name + r'.csv'
    weather_data_per_camera_agg.to_csv(weather_data_per_camera_agg_path,index=False)
    #print (i)                

### Some debug steps

camera_station_mapping3['weather_file_name'] = camera_station_mapping3.stid + ".csv"

available_station_in_mapping = set(camera_station_mapping3.weather_file_name.unique())

camera_station_mapping4.head()

os.chdir(weather_data_file_path + r"\split_files\ByWeatherStation")
available_station_in_file = set(os.listdir())

missing_file = list (available_station_in_mapping - available_station_in_file)

missing_file 

camera_station_mapping4_missing_weather = camera_station_mapping4[camera_station_mapping4['weather_file_name'].isin(missing_file)]
camera_station_mapping4_missing_weather

camera_station_mapping4_missing_weather['properties.description.id'].unique()

camera_station_mapping5 = camera_station_mapping3[~camera_station_mapping3['stid'].isin(weather_station_data_available)]


camera_station_mapping5.head()

camera_station_mapping5.stid.unique()
